app_name/requirements/base.txt
attrs==22.2.0
click==8.1.3
Django==4.2.9
django-appconf==1.0.5
django-auth-ldap==4.1.0
django-compressor==4.3.1
django-rq==2.7.0
EasyProcess==1.1
Pillow==9.4.0
numpy==1.24.2
python-ldap==3.4.3
pytz==2022.7.1
rcssmin==1.1.1
redis==4.5.1
rjsmin==1.2.1
requests==2.28.2
sqlparse==0.4.3
django-sendfile2==0.7.0
dj-pagination==2.5.0
python-logstash-async==2.5.0
rules==3.3
coreapi==2.3.3
django-filter==22.1
Markdown==3.4.1
djangorestframework==3.14.0
Pygments==2.14.0
drf-spectacular[sidecar]==0.26.0
Shapely==2.0.1
opencv-python-headless==4.7.0.72
h5py
django-cors-headers==3.14.0
furl==2.1.3
natsort==8.3.1
django_cleanup==7.0.0
python-dotenv==1.0.0
onnxruntime==1.14.1
imutils==0.5.4
scipy==1.10.1
scikit-learn==1.2.2
gdown==4.6.4
funcy==1.18
pandas==1.5.3
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
django-allauth==0.53.0
celery==5.2.7
django-celery-beat==2.5.0
flower==1.2.0
uritemplate==3.0.1
protobuf==3.20.3
PyYAML
kfp==1.8.19
hiredis==2.2.2
dj-rest-auth==5.1.0
django-allauth[socialaccount]==64.1.0